+++ src/driftguard/collectors/__init__.py
@@ -3,6 +3,7 @@
 from driftguard.collectors.base import BaseCollector
 from driftguard.collectors.csv_collector import CsvCollector
 from driftguard.collectors.json_collector import JsonSchemaCollector, OpenApiCollector
+from driftguard.collectors.json_sample_collector import JsonSampleCollector
 from driftguard.collectors.mysql_collector import MysqlCollector
 from driftguard.collectors.postgres_collector import PostgresCollector
 from driftguard.collectors.sqlite_collector import SqliteCollector
@@ -10,6 +11,7 @@ from driftguard.collectors.sqlite_collector import SqliteCollector
 __all__ = [
     "BaseCollector",
     "CsvCollector",
+    "JsonSampleCollector",
     "JsonSchemaCollector",
     "MysqlCollector",
     "OpenApiCollector",